Refactor rendering & text: replace EmbeddedText
Replace the old EmbeddedText system with direct drawing helpers and simplify rendering math. Key changes:

- TextEntity: remove EmbeddedText usage, manage displayed string/scale locally, update movement and collision math, add onRender to draw text and extra sprite.
- Entity: remove EmbeddedText inner class and embedTexts list; add drawText and drawSprite helpers; change position fields to int; simplify setSprite and render to use Graphics2D copies.
- Scene: return int dimensions/centers instead of double.
- EngineConfig: bump version to [BETA] v0.1.2.

Breaking changes: EmbeddedText API and double-based x/y coordinates removed—callers must be updated.

+++ src/org/quirkle/resourceEngine/Scene.java
@@ -8,10 +8,10 @@ import java.util.concurrent.CopyOnWriteArrayList;
 public class Scene {
     protected final List<Entity> sceneEntities = new CopyOnWriteArrayList<>();
 
-    public double getWidth() { return EngineConfig.WINDOW_WIDTH; }
-    public double getHeight() { return EngineConfig.WINDOW_HEIGHT; }
-    public double getCenterX() { return getWidth() / 2.0; }
-    public double getCenterY() { return getHeight() / 2.0; }
+    public int getWidth() { return EngineConfig.WINDOW_WIDTH; }
+    public int getHeight() { return EngineConfig.WINDOW_HEIGHT; }
+    public int getCenterX() { return (int) (getWidth() / 2.0); }
+    public int getCenterY() { return (int) (getHeight() / 2.0); }
 
     public Scene() {
         onCreate();
